Header and Title components splitter. Logout mutation implemented. Working on better UI flow after logout user
