+++ src/graphql/mutations.js
@@ -30,3 +30,9 @@ export const UPDATE_USER = gql`
     }
   }
 `
+
+export const LOGOUT = gql`
+  mutation {
+    logout
+  }
+`
